OXDEV-9063 Added dry-run option for move/delete

+++ src/ImageManager/Service/ImageManagerService.php
@@ -16,8 +16,10 @@ use Psr\Log\LoggerInterface as PsrLoggerInterface;
 class ImageManagerService implements ImageManagerServiceInterface
 {
     private const IMAGE_MOVED_SUCCESSFUL = 'Moved image: %s to %s';
+    private const IMAGE_MOVED_DRY_RUN = '[DRY-RUN] Move %s to %s';
     private const IMAGE_MOVED_FAILED = 'Failed to move image: %s. Error: %s';
     private const IMAGE_DELETED_SUCCESSFUL = 'Deleted image: %s';
+    private const IMAGE_DELETED_DRY_RUN = '[DRY-RUN] Delete %s';
     private const IMAGE_DELETED_FAILED = 'Failed to delete image: %s';
 
     public function __construct(
@@ -25,31 +27,39 @@ class ImageManagerService implements ImageManagerServiceInterface
         private readonly PsrLoggerInterface $logger
     ) {
     }
-    public function moveImages(ImageCollectionInterface $images, string $destination): void
+    public function moveImages(ImageCollectionInterface $images, string $destination, bool $dryRun = false): void
     {
         foreach ($images->getAll() as $image) {
             $sourcePath = rtrim($image->getDirectory(), '/') . '/' . $image->getImageName();
             $destinationPath = rtrim($destination, '/') . '/' . $image->getImageName();
 
-            try {
-                $this->fileSystemUtils->moveFile($sourcePath, $destinationPath);
-                $this->logger->info(sprintf(self::IMAGE_MOVED_SUCCESSFUL, $sourcePath, $destinationPath));
-            } catch (\Exception $e) {
-                $this->logger->error(sprintf(self::IMAGE_MOVED_FAILED, $sourcePath, $e->getMessage()));
+            if ($dryRun) {
+                $this->logger->info(sprintf(self::IMAGE_MOVED_DRY_RUN, $sourcePath, $destinationPath));
+            } else {
+                try {
+                    $this->fileSystemUtils->moveFile($sourcePath, $destinationPath);
+                    $this->logger->info(sprintf(self::IMAGE_MOVED_SUCCESSFUL, $sourcePath, $destinationPath));
+                } catch (\Exception $e) {
+                    $this->logger->error(sprintf(self::IMAGE_MOVED_FAILED, $sourcePath, $e->getMessage()));
+                }
             }
         }
     }
 
-    public function deleteImages(ImageCollectionInterface $images): void
+    public function deleteImages(ImageCollectionInterface $images, bool $dryRun = false): void
     {
         foreach ($images->getAll() as $image) {
             $filePath = rtrim($image->getDirectory(), '/') . '/' . $image->getImageName();
 
-            try {
-                $this->fileSystemUtils->deleteFile($filePath);
-                $this->logger->info(sprintf(self::IMAGE_DELETED_SUCCESSFUL, $filePath));
-            } catch (\Exception $e) {
-                $this->logger->error(sprintf(self::IMAGE_DELETED_FAILED, $filePath, $e->getMessage()));
+            if ($dryRun) {
+                $this->logger->info(sprintf(self::IMAGE_DELETED_DRY_RUN, $filePath));
+            } else {
+                try {
+                    $this->fileSystemUtils->deleteFile($filePath);
+                    $this->logger->info(sprintf(self::IMAGE_DELETED_SUCCESSFUL, $filePath));
+                } catch (\Exception $e) {
+                    $this->logger->error(sprintf(self::IMAGE_DELETED_FAILED, $filePath, $e->getMessage()));
+                }
             }
         }
     }
